create new separate class for UIScreen extension
