+++ FTPopOverMenu_Swift/FTPopOverMenu.swift
@@ -603,14 +603,3 @@ class FTPopOverMenuCell: UITableViewCell {
     }
 
 }
-
-extension UIScreen {
-    
-    public static func ft_width() -> CGFloat {
-        return self.main.bounds.size.width
-    }
-    public static func ft_height() -> CGFloat {
-        return self.main.bounds.size.height
-    }
-    
-}
